+++ integration/e2e/e2e_test.go
@@ -98,16 +98,14 @@ func TestAllInOne(t *testing.T) {
 			// query an in-memory trace
 			queryAndAssertTrace(t, apiClient, info)
 
-			// search an in-memory trace
-			util.SearchAndAssertTrace(t, apiClient, info)
-
 			// wait trace_idle_time and ensure trace is created in ingester
 			require.NoError(t, tempo.WaitSumMetricsWithOptions(e2e.Less(3), []string{"tempo_ingester_traces_created_total"}, e2e.WaitMissingMetrics))
 
 			// flush trace to backend
 			callFlush(t, tempo)
 
-			// traceql search and verify trace is searchable once cut
+			// search for trace in backend
+			util.SearchAndAssertTrace(t, apiClient, info)
 			util.SearchTraceQLAndAssertTrace(t, apiClient, info)
 
 			// sleep
@@ -251,9 +249,6 @@ func TestMicroservicesWithKVStores(t *testing.T) {
 			// query an in-memory trace
 			queryAndAssertTrace(t, apiClient, info)
 
-			// search an in-memory trace
-			util.SearchAndAssertTrace(t, apiClient, info)
-
 			// wait trace_idle_time and ensure trace is created in ingester
 			require.NoError(t, tempoIngester1.WaitSumMetricsWithOptions(e2e.Less(3), []string{"tempo_ingester_traces_created_total"}, e2e.WaitMissingMetrics))
 			require.NoError(t, tempoIngester2.WaitSumMetricsWithOptions(e2e.Less(3), []string{"tempo_ingester_traces_created_total"}, e2e.WaitMissingMetrics))
@@ -264,6 +259,10 @@ func TestMicroservicesWithKVStores(t *testing.T) {
 			callFlush(t, tempoIngester2)
 			callFlush(t, tempoIngester3)
 
+			// search for trace
+			util.SearchAndAssertTrace(t, apiClient, info)
+			util.SearchTraceQLAndAssertTrace(t, apiClient, info)
+
 			// sleep for one maintenance cycle
 			time.Sleep(5 * time.Second)
 
@@ -272,7 +271,7 @@ func TestMicroservicesWithKVStores(t *testing.T) {
 				require.NoError(t, i.WaitSumMetrics(e2e.Equals(1), "tempo_ingester_blocks_flushed_total"))
 			}
 			require.NoError(t, tempoQuerier.WaitSumMetrics(e2e.Equals(3), "tempodb_blocklist_length"))
-			require.NoError(t, tempoQueryFrontend.WaitSumMetrics(e2e.Equals(3), "tempo_query_frontend_queries_total"))
+			require.NoError(t, tempoQueryFrontend.WaitSumMetrics(e2e.Equals(4), "tempo_query_frontend_queries_total"))
 
 			// query trace - should fetch from backend
 			queryAndAssertTrace(t, apiClient, info)
@@ -287,15 +286,19 @@ func TestMicroservicesWithKVStores(t *testing.T) {
 			info = tempoUtil.NewTraceInfo(time.Now(), "")
 			require.NoError(t, info.EmitAllBatches(c))
 
-			// query an in-memory trace
+			// query by id
 			queryAndAssertTrace(t, apiClient, info)
 
-			// search an in-memory trace
-			util.SearchAndAssertTrace(t, apiClient, info)
+			// wait trace_idle_time and ensure trace is created in ingester
+			require.NoError(t, tempoIngester1.WaitSumMetricsWithOptions(e2e.Less(4), []string{"tempo_ingester_traces_created_total"}, e2e.WaitMissingMetrics))
+			require.NoError(t, tempoIngester3.WaitSumMetricsWithOptions(e2e.Less(4), []string{"tempo_ingester_traces_created_total"}, e2e.WaitMissingMetrics))
 
-			// search the backend. this works b/c we're passing a start/end AND setting query ingesters within min/max to 0
-			now := time.Now()
-			util.SearchAndAssertTraceBackend(t, apiClient, info, now.Add(-20*time.Minute).Unix(), now.Unix())
+			// flush trace to backend
+			callFlush(t, tempoIngester1)
+			callFlush(t, tempoIngester3)
+
+			// search for trace
+			util.SearchAndAssertTrace(t, apiClient, info)
 
 			// stop another ingester and confirm things fail
 			err = tempoIngester1.Kill()