+++ pkg/search/util.go
@@ -1,20 +1,27 @@
 package search
 
 import (
-	"github.com/grafana/tempo/pkg/model/trace"
-	"github.com/grafana/tempo/pkg/tempofb"
 	"github.com/grafana/tempo/pkg/tempopb"
 	"github.com/grafana/tempo/pkg/traceql"
-	"github.com/grafana/tempo/pkg/util"
+)
+
+const (
+	ErrorTag        = "error"
+	StatusCodeTag   = "status.code"
+	StatusCodeUnset = "unset"
+	StatusCodeOK    = "ok"
+	StatusCodeError = "error"
+
+	RootSpanNotYetReceivedText = "<root span not yet received>"
 )
 
 func GetVirtualTagValues(tagName string) []string {
 	switch tagName {
 
-	case trace.StatusCodeTag:
-		return []string{trace.StatusCodeUnset, trace.StatusCodeOK, trace.StatusCodeError}
+	case StatusCodeTag:
+		return []string{StatusCodeUnset, StatusCodeOK, StatusCodeError}
 
-	case trace.ErrorTag:
+	case ErrorTag:
 		return []string{"true"}
 	}
 
@@ -33,16 +40,6 @@ func GetVirtualTagValuesV2(tagName string) []tempopb.TagValue {
 	}
 
 	return nil
-}
-
-func GetSearchResultFromData(s *tempofb.SearchEntry) *tempopb.TraceSearchMetadata {
-	return &tempopb.TraceSearchMetadata{
-		TraceID:           util.TraceIDToHexString(s.Id()),
-		RootServiceName:   s.Get(trace.RootServiceNameTag),
-		RootTraceName:     s.Get(trace.RootSpanNameTag),
-		StartTimeUnixNano: s.StartTimeUnixNano(),
-		DurationMs:        uint32((s.EndTimeUnixNano() - s.StartTimeUnixNano()) / 1_000_000),
-	}
 }
 
 // CombineSearchResults overlays the incoming search result with the existing result. This is required